move type to own file
