+++ src/views/list/types.ts
@@ -1,0 +1,5 @@
+import { ListMediaMedia } from '@/graphql/types'
+
+export type ListMedia = {
+  [key: number]: { media: ListMediaMedia | null; loading: boolean } | undefined
+}
